backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
scikit-learn
numpy
pydantic
python-multipart